Move _custom logic from converter to API
In preparation for XtractHub running conversion logic, move the logic to format the custom block into the API submission processing. Not only is this change more in line with centralizing the preprocessing in API, it removes the only non-parsing logic form the converter, preparing for complete converter replacement with XtractHub.