Remove source_info requirement from Validator, add variant Globus link support
source_info requirement in Validator is redundant - source_id and source_name should already be present in dataset entry before validation, and this removes requirement for mdfcs.utils in XtractHub. Variant of Globus Web App link spotted in wild, support added. Config field made for GWA links for easier expansion in the future if/when new link formats are discovered.